Increased customisability by increasing use of interfaces, while moving logic out of genotypes
