+++ README.md
@@ -11,6 +11,17 @@ Some Key Features:
 ## Documentation
 The documentation is stored on the [GoEvo Wiki](https://github.com/JoshPattman/goevo/wiki).
 
+## Overview of Structure
+The parent module, `goevo`, defines a set of interfaces that lay out the behavior for which the different components must implement. However, the parent module does not implement any of these interfaces, but instead, some default implementations are stored in the sub-modules. As long as you can provide implementations for the following interfaces, it is possible to use any combination of implementations to create a customized evolutionary simulation:
+
+- `Population[T]`: Represents a population storing a set of genotypes of type `T`. All default implementations of this interface are stored in `pop/`.
+- `MutationStrategy[T]`: Represents a mutation strategy that can be applied to a genotype of type `T` to perform a mutation. The default implementations reside in `geno/`. You can have multiple mutation strategies for any given genotype type.
+- `CrossoverStrategy[T]`: Represents a crossover strategy that can be applied to two genotypes of type `T` to perform a crossover (this can have any number of parents). The default implementations reside in `geno/`. You can have multiple crossover strategies for any given genotype type.
+- `Selection[T]`: Represents a selection strategy that can be applied to a population of genotypes of type `T` to select a subset of the population. The default implementations reside in `selec/`. You can have multiple selection strategies for any given genotype type.
+- Genotypes: These do not need to implement any interfaces themselves. However, they are the DNA upon which the crossover and mutation strategies are applied. The default implementations reside in `geno/`.
+
+It is also entirely possible to only use certain components of the package. For example, a continuous evolutionary simulation may not require a `Selection` or `Population`, as these are inherited from interactions with the environment.
+
 ## Built-In Features List
 ### Algorithms
 - [NEAT](https://nn.cs.utexas.edu/downloads/papers/stanley.ec02.pdf) (Neuro Evolution of Augmenting Topologies)
