+++ pop/hillclimber/population.go
@@ -8,11 +8,11 @@ var _ goevo.Population[int] = &Population[int]{}
 type Population[T any] struct {
 	A            *goevo.Agent[T]
 	B            *goevo.Agent[T]
-	Selection    goevo.Selection[T]
+	Selection    goevo.SelectionStrategy[T]
 	Reproduction goevo.Reproduction[T]
 }
 
-func NewPopulation[T any](initialA, initialB T, selection goevo.Selection[T], reproduction goevo.Reproduction[T]) *Population[T] {
+func NewPopulation[T any](initialA, initialB T, selection goevo.SelectionStrategy[T], reproduction goevo.Reproduction[T]) *Population[T] {
 	return &Population[T]{
 		A:            goevo.NewAgent(initialA),
 		B:            goevo.NewAgent(initialB),
