+++ pop/simple/population.go
@@ -9,12 +9,12 @@ var _ goevo.Population[int] = &Population[int]{}
 // Population has a single species, and generates the entire next generation by selcting and breeding from the previous one.
 type Population[T any] struct {
 	Agents       []*goevo.Agent[T]
-	Selection    goevo.Selection[T]
+	Selection    goevo.SelectionStrategy[T]
 	Reproduction goevo.Reproduction[T]
 }
 
 // NewPopulation creates a new SimplePopulation with n agents, each with a new genotype created by newGenotype.
-func NewPopulation[T any](newGenotype func() T, n int, selection goevo.Selection[T], reproduction goevo.Reproduction[T]) *Population[T] {
+func NewPopulation[T any](newGenotype func() T, n int, selection goevo.SelectionStrategy[T], reproduction goevo.Reproduction[T]) *Population[T] {
 	agents := make([]*goevo.Agent[T], n)
 	for i := range agents {
 		agents[i] = goevo.NewAgent(newGenotype())
